Move functions in ResOpsES to the repository
